Added function binding
Binding is implicitly used with prototypical method calls:

    let t = [msg: 'hi', doit: fn(t, f) -> f(t.msg)]
    t.doit(t, print)       # no binding
    bind(t.doit, t)(print) # binding
    t->doit(print)         # implicit binding
    call(t->doit, print)   # binding not necessarily only invocation

Unfortunately, binding in the current function implementation is a bit
clunky since sbfn calls and fcalls have different interfaces. This
causes an uneeded allocation when a tail call could be possible.

The more important optimization involves removing the actual bind call
in method invocation, which should be much more common. Of course this
also ended up being clunky with the current parser.