Cleaned up parser
This includes much refactoring that focused on cleaning code rather
than performance. Still, the parsing is relatively complex.

The lexical analysis stages have been moved entirely into parse.c.
While this does make parse.c excessively large, it hides the lexing
mechanics from users, which was growing increasingly entangled with
the parser. Now all parsing/lexing/encoding structures are hidding
behind a handful of simple parse functions.

Another big change was switching from a simple enum for representing
to using bit positions. Each token represents a filled bit position.
This has the neat property and allowing words with multiple bits set to
represent sets of multiple tokens. These sets can then be matched
against in constant time. The biggest downside of this approach is that
it's now pretty much impossible for switch statements to compile down
to jump tables, but considering the sparsity of conditions in the parsing
code, I suspect this isn't too much of a loss.

In an effort to better organize the parser, most chunks of related
information was grouped into structs specific to stages of the grammar
that are allocated on the stack when parsed. This keeps that stage's
specific information from leaking into nested stages. In the end, the
parsing struct became so small, that I just decided to allocate
everything on the stack. This, combined with passing pointers to
everything does mean that there is probably a problematic amount of
stack usage, but I figure that can be looked at if it becomes an issue.

Interestingly, almost all allocations were removed in the parser.
Structs are allocated on the stack, jumps are patched through reading
the emitted bytecode, and everying is encoded in one pass so only just
the amount of data needed is allocated.

Finally, the rest of the grammar has been implemented. There's still
quite a few things that need tweaking, such as potentially adding
type definitions and I'll probably remove for loop expressions since
they create problematic confusion when mixed with frames.

Man this set of changes was exhausting.