Tweaked the parser a bit
Decided to remove for loop expressions. They weren't that difficult
to parse, and they are kinda neat as list-comprehension like constructs.
However, their syntax became confusing when interacting with frames.

    return for (k,v = pairs(table)) if(isprime(k)) 2*k, v

Returns a single element iterator and whatever was previously in v.

Even if wrapped in parens to remove that surprising ambiguity it
would require support at the frame level which would need quite
a bit of changes in the parser.
    return (for (k,v = pairs(table)) if(isprime(k)) 2*k, v)

In the end, I suspect for-expressions will not be worth the extra
complexity or confusion and can be replaced by their
grammar-independant and less ambigous map/filter counterparts.
(Which they already compile as calls to).
    return map((fn(k,v) -> 2*k,v),
            filter(fn(k,v) -> isprime(k), pairs(table)))

Unrelatedly, lookaheads now use the full token. This is less
efficient and requires an annoying lex-struct allocation, but
fixes issues with grabbing tokens too soon.
    a == b => a = (=b)