Fixed issues with leading/trailing whitespace in blocks
Handling whitespace turned out to be relatively complicated in blocks.

Basically, in a block, if the indention ever returns to the original
level, then the block is ended. Brackets also change to block level,
but don't change the indention level, making brackets implicitly
necessary to match brackets. The final rule is that the root block
is special, and isn't ended by indention.

Now it's possible to do this:

    fn hello()
         1
        1
         1
          1
         1
        return 'tada'