Implemented core builtins
Like all of the current parts of the language, this will
probably be in flux for a while, but in the commit is the
initial framewrk for the set of builtins in the library.

The whole purpose of Mu is to provide a simple scripting
language with just enough functionality that the user can write
powerful scripts without _needing_ other libraries. The set of
builtins was selected to be powerful and small. This goal led to
some difficult design decisions on what to include.

Implemented:
- Type functions (num, str, tbl, fn_)
- Comparison operations (==, !=, is, <, <=, >, >=)
- Arithmetic operations (+, -, *, /, abs, floor, ceil, //, %, ^, log)
- Trignometry operations (cos, acos, sin, asin, tan, atan)
- Bitwise/Set operations (&, |, ~, &~, <<, >>)
- String representation (parse, repr, bin, oct, hex)
- Data structure operations (len, const, ++, sub)
- Table operations (tail, push, pop)
- String operations (find, replace, split, join, pad, strip)
- Function operations (bind, comp, map, filter, reduce, any, all)
- Iterator operations (iter, pairs, range, repeat, zip, chain, take, drop)
- Ordering operations (min, max, reverse, sort)

Not Implemented:
(May require OS support, haven't really decided how to handle yet)
- Random operations (random, seed)
- Time operations (time, sleep, call_in, call_every)
- IO operations (print, error)
- File operations (read, write, append, pipe)