Changed to variable-length bytecode for 16-bit immediates
After toying around with several Mu scripts, started hitting the upper
limits of the existing bytecode. It seems that 8-bit immediates are
simply not sufficient for jumps and loading variables and breaks
down in moderately large loops and large declarations of tables.

A quick solution is to up the limit to 16-bits. Since this results
in 3-byte ops, there's very little benefit to keeping consistently
sized opcodes, especially since space is a concern. After switching
to variable-length bytecode, the increased instruction size results
in a roughly 12% increase in the size of the generated code struct.