Changed Mu constant declarations in C to be simpler and more portable
Before, constant declarations could be done in expression contexts using
the mc* set of macros. This worked pretty well, except for a few problems:
- Unintuitive, why is declaring a constant in a function-like macro?
- Confusing, internally the arguments to the macro were expanded in
  declaration contexts and could lead to confusing error when a
  non-constant was passed in. They were also kinda hard to explain.
- Limited, since they could only generate expression level constructs,
  the macros couldn't take advantage of gcc's constructors or be
  expanded to much else and initialization had to be done with flags.
- Non-portable, the mc* set of macros relied entirely on gcc's
  statement-expressions. Even then the implementation was a bit hacky
  and relied on static const declarations in the statement-expressions.

The new constant declarations use all-caps versions of their runtime
counterparts:

 #define STRING_CONSTANT string_constant()
 mu_pure mu_t string_constant(void)

 MSTR(string_constant, "string constant")

This version gives the constants a chance to actually be declared at the
file level without any expression workarounds. It generates a pure
function that can perform and remaining runtime work. Unfortunately, the
declaration of the constant is a bit messy, requiring the full function
declaration and an define for the evaluated constant which isn't very
macroizable.

A neat side-effect of this switch, is that the generated function
can be used instead of the actual variable for being stored in
table constants, allowing the full description of the table to be stored
in rom.