Changed to variable-length immediates in bytecode
Since space is a primary concern for Mu, especially in dynamically
allocated memory, generating small bytecode is beneficial. By using
variable-length immediates (with a simply vlq encoding), the bytecode
has both the benefits of compressability and the full range of immediates
possible, at the cost of additional runtime operations.

Another nice benefit is that the bytecode is completely portable between
32-bit/64-bit Mu as long as the other limitations in 32-bits are
followed (only 16-bit imms/fns/tables).

Unfortunately, the parser currently can't emit variable length jumps to
take advantage of the new encoding. Currently a static size is emitted,
but the vm can execute any size. This is to allow portable execution in
case a future implementation works around this limitation.