Added buffers for C internals
Buffers are simply mutable blobs of memory that can be used in C.
They are more efficient for mutation than strs, but aren't exposed
the the Mu side, since every usage could be replaced by strs.

Buffers replace the previous concept of mstrs, in that they are
directly stored in the intern table and support amortized constant
growth for easy string creation. However, their purpose is expanded
to allow C to pass data directly through Mu if it doesn't need to be
garbage collected.