Removed main program
Mu is intended to be a simple, embeddable scripting language
with the goal of being easy to integrate into C programs. While
it makes it easier to try out the language, Mu isn't really
intended to be used as a standalone program.

The main program will be separated out into a standalone repository
dubbed mux, the Mu executable. As a separate project, mux will
also serve as a simple example of how to integrate Mu into a project.