Reverted arg order change in find/replace
Initially to match existing libraries and open up easy
implementation of regex equivalent, the match-first order
is generally non-intuitive with the rest of the functional
string builtins.

With the string being searched in as the target for the
find/replace, using the match-second order makes more sense.

This does make a regex library with regex objects less
compatible, but ease of use wins this compromise.