Moved to less-granular 16-bit instructions
Prior to this patch, Mu used a variable-length 8-bit instruction
set with full vlq-like encoding for immediate values. Although
a very flexible format, the instructions themselves didn't take
advantage of the compressibility, with mostly 16-bit or 32-bit
instructions for jumps due to limitations in the code generation.

This patch adopts a much simpler immediate encoding. All immediates
start as 8-bits, and if the value is 0xff, the full 16-bit immediate
is appended to the instruction.

before:

    f014        jump 20
    f08148      jump 200
    f08f50      jump 2000

after:

    f014        jump 20
    f0c8        jump 200
    f0ff07d0    jump 2000

The simpler encoding reduces the complexity of decoding instructions
to a single branch, is actually smaller for functions with ~200
immediates, and plays nicer with the branch generation.