Added weak functions to avoid inherit reference cycles
Also adopted flags for structure attributes to allow additional
attributes without space increase.

Mu relies on reference counting and until a cycle-collector is
added, any cycles are prone to memory leakage. The initial rule
will be to avoid cycles unless logic is in place to destruct any
cycles before collection.

Unfortunately, because of the simplicity of Mu's scoping and code
generation, the following construct generates an implicit reference
cycle between the function declaration and it's own scope:

    fn hi()
        print('hi')

This patch adds weak functions, a quick hack to allow the compiler
to mark `hi` as weak and allow function declarations to be
collected.

This is unlikely to be a long-term solution.