Fixed trailing padding in sizeof calculation for vl structures
It was an interesting discovery to find out that the alignment of
a structure has potential to change its size. In C, the size of the
structure also matches its alignment, to avoid complications when
used with arrays/buffers calculated from its size.

For the buffer and string classes, the data stored after the structure
was expected to fill this unused padding, but the current math actually
puts the data after the padding, wasting several words with every
buffer/string.

The best solution found was to adopt flexible array members and add
the offsetof builtin wrapped in mu_offsetof for portability concerns.
This has the additional benefit of simplifying member access.