Adopted compact tables with insertion ordering
Based on the relatively recent innovation of python's compact
dictionaries, this patch adopts compact tables. Compact tables
provide a decent ram savings, as well as provides an inherent
ordering of insertions.

The basic idea is that the hash tables are split into two seperate
arrays. One for the entry contents, and a sparse hash array containing
indirect pointers to the entries. This indirect hash array can use the
smallest integer type needed for addressing all entries, giving the data
structure an overall win in terms of memory consumption. And as a nice
side effect, the backing array maintains insertion order.

For an abstract example, a non-compact table containing 5 entries:
['a':'b', nil:nil, nil:nil, 'c':'d', nil:nil, 'g':'h', 'e':'f', nil:nil]

vs a compact table containing 5 entries:
[1, 0, 0, 2, 0, 4, 3, 0] ['a':'b', 'c':'d', 'e':'f', 'g':'h']

If we assume 32-bit mu types and that the hash references only need 1
byte, the first uses 64 bytes, while the second uses only 40 bytes.

---

Unfortunately, this patch wasn't as simple as tossing in the second hash
array and calling it good. The bounded table header (4 words) caused a
lot of problems. It couldn't fit the additional pointer for the hash array
or any sort of subdivision for the entry array.

In the end, I crafted a rather questionable formula to subdivide the
entry array to make space for the hash array:
slots = npw2(ceil((pairsize*len) / (pairsize - slotsize)))

This formula gives this interesting constraint when adjusted for storage
size (I did play with other multipliers, but x1 seemed to work the best):
slots = slots + entries

What all this math (only calculated at resize time) means, is that the
size of the hash array and entry array are the same, and can be stored
as a single length. Also, the percentage of the entry array used for
entries is equal to the load factor of the hash array. This is enough to
fit everything into the table header.

An odd side effect of this formula is that the load factor decreases as
the size of the table increases. For small tables this is great, since
small tables need speed less than larger tables. But in the worst case,
the largest tables (%12) end up with a load factor of 0.5. This might be
a problem later, but with the wins for small tables, I'm not going to
lose sleep over it for now.

mu32 worst-case load factor calculations:
       entries  slotsize  load factor
          <240         1        0.875
       <57,344         2        0.75
       <65,535         4        0.5

mu64 worst-case load factor calculations:
       entries  slotsize  load factor
          <240         1       0.9375
       <57,344         2       0.875
<3,221,225,472         4       0.75
<4,294,967,295         8       0.5

---

I threw together a quick test to measure the memory consumption (note I
didn't remove the baseline, so the lower numbers are a bit skewed, but do
show the the overall win)

command:
valgrind mu -e 'print(len(tbl(zip(range(-1000000), repeat("a")))))'

results (entries vs ram in bytes):
entries  non-compact      compact
      1       44,502       12,054
     10       47,831       15,383
    100       84,312       47,768
   1000      429,657      397,209
  10000    3,768,410    3,735,962
 100000   40,432,731   40,400,283
1000000  387,152,988  387,120,540