Modified push/pop/sub to take advantage of table ordering
The adoption of table ordering as a feature is probably the biggest
divergence from python that mu is taking in the adoption of compact
tables.

Before compact tables, it was expensive to maintain table ordering.
And there is the possibility that a new, faster/smaller data structure
will be invented for unordered key-value stores. For these reasons,
the order of dictionaries remains undefined in python.

However, the unordered nature of tables can be a surprise to new users
unfamiliar with how hash tables work internally. Just search for
"language" "name of table type" "order" and you will find many results
of users asking why their tables are out of order.

(to be fair, a lot of these results are about parameters to functions,
and python _did_ decide that keyword arguments are going to be ordered)

Since arbitrary order simply does not match the expecation of new users,
and an efficient ordered table exists (compact tables), mu is going to
adopt table ordering whole-heartedly.

---

What was this pr about again? oh, the ordering of tables also provides
a bit more reasonable of an implementation of push/pop/sub on tables
with non-integer keys.

Now push/pop/sub operate on table order as though it was an array. And
since it is internally, this allows the operations to no longer have to
allocate extra space while it moves the keys around. Neat.