Capitalized constants
Capitalized
e -> E
pi -> PI

Unchanged
inf
nil
true
false

Decided to leave anything that could be construed as a language keyword
in lower case. Some of these (true, inf) are just constants, but their
use is not really as magic numbers. (although perhaps true is too
strange in mu).

This should be enough to encourage users to follow the same style in
their code. Capitalized constants is a nice standard for languages
with capitalization.