Added tail-chain tables to bufs
This is a big change, since before tables were the only types that
allowed lookups. But it is very beneficial for C libraries that want
to represent opaque data in mu, since they can now provide inheritance
of method-like functions.

This also tightly couples the idea that bufs are a more efficient
C equivalent of tables, since they are almost a complete replacement.

---

Interesting thing to note, a buf never actually needs to participate
in tail chains, since itself does not actually host the lookups. When
extending a table by its tail, it's just the tail that is used.

Also the layout of bufs requires a lot of arithmetic to access the dtor
and buf, since these are stored at the end. This is beneficial for
catching buffer overflows, but it may be worth rearranging these for
more efficient lookups (The dtor+tail is 3-bit aligned hint hint).