Increased even distribution of table's hash function
I've had some concern for a while of the hash function distribution,
since the integer representations of floats and pointers is very
unlikely to be evenly distributed. Unfortunately, hash functions that
can fix this distribution are relatively complicated, and involved
many operations that may be more costly for small tables than the
benefit for large tables.

There is Knuth's multiplicative hash, which just involves multiplication
by an irrational fraction of the max int. But it has a big flaw in that
multiplication only propogates entropy upwards, and later masking for
the lookup in the table will drop the most influenced bits.

However, masking doesn't need to occur from the bottom. Instead, we can
just shift the integer down, intentionally keeping the most influenced
bits.

As an added benefit, this is will put the most influenced bits on the
high side, allowing the simple iteration during hash conflicts to scan
a hopefully empty region of the hash table.