Added back read-only tables
Read-only tables were removed a while back because of the limited
benefit and implementation costs. However, they are needed to allow
mu to store tables in ROM.

Now that the codebase is more concrete (and the type tag is no longer
saturated), we can add read only tables back in.

It does come with an immediate ROM cost, but should RAM cost at runtime.