Added scope checking
Now, if a variable is used are assigned before it is declared in a let
statement or equivalent, it will error at compile time.

This actually turned out a lot simpler a task than what I was prepared
for. Since the parser tries to tie as many leaves in the grammar
together as possible (to save on code space), there was only a couple
of rules that needed to check scoping. Added a table with tail chain
for the current scope check, and it was done.

An interesting observation of the grammar is that assignment doesn't
actually matter in terms of scope checking, only if it is a let
statement. This is reflected in the single condition for the grammar
rule.

As an added benefit, now we can actually get the local count correct
for the scope table created in the vm.

---

As a side note, this also opens up the possibility to inline builtins
that passed as the scope during compile time. Implementing a blunt
inline for all of the scope was rather easy and seemed to work quite
well, with the big downside that assignment to builtins would no longer
work. However, this was ok as the builtins were usually readonly (for
storing in ROM).

Unless we are writing a repl. The repl calls compile each iteration with
the current scope as an argument. But the current scope must be mutable
for the operations to make sense.

The end conclusion I came up with, is that the compiler can inline
builtins, but only if it knows they are readonly. The table should does
carry this information, but unfortunately there is no API to look this
up in an efficient manner.

Measuring the runtime only gave somewhat dissapointing results (although
it was an improvement), so I decided to table to problem for now.