Fixed assignment of readonly table in tail chains
The behaviour of assigning to a table chain that contains readonly
tables is a bit subtle.

I would expect this to work:

    let a = const([a:1])
    let b = tbl([b:1], a)
    b.b = 2

But this should not:

    let a = [a:1]                # not const
    let b = const(tbl([b:1], a)) # const
    let c = tbl([c:1], b)        # not const
    c.a = 2

It turns out I had implemented the readonly table assignment wrong
twice. The current implementation fails on the first case, and the
old implementation would assign 2 to a new local slot in c.