Restructured mu to fit in a single repo
The core files are still contained in a single mu directory, and
I am planning to keep a subtree branch for only that directory, to
make is easier to embed in projects.

The root of the project can now contain other project metadata, such
as makefiles, tests, docs, other modules.