Added the dis module for dissasembling mu objects
The dis module is a heavy-handed tool for debugging. It dumps
all the information it knows about the object passed to it, bypassing
any notions of oop encapsulation.

It's useful, but also comes with a hefty code cost. It should be built
when debugging Mu itself, but should probably be left out of external
projects.