Added DEF checks for null, allowing conditional table entries
The biggest pain with the current approach for the DEF macros is
the annoying fact that C doesn't allow preprocessor statements in
macro arguments.

I considered alternative designs for the DEF macros. But few fit
cleanly in C and each came with their own set of hacks. At the end
of the day I stayed with the design the fit best into the C language.

This leaves the option for the user to do conditional logic outside
of the DEF macro, and expand to nil if a table entry isn't included.
This did require a few changes to let nil entries be provided to the
DEF macros.

Fortunately, none of this additional logic will matter if a Mu
preprocessor is written for the DEF macros. (forshadowing?)