Added the repl library
The goal of the repl library is to provide a lightweight interpreter
that can be slapped on projects using mu to provide an easy method of
interacting with the project. The benefits are quick debugging and the
ability to look inside the project while it is running. The repl can
also act as a packaged alternative to a more traditional handwritten
shell for commands.

The repl tries to lean toward the simplier lightweight side of repl
implementations, but does come with history and table-based
tab-completion.

The implementation is based off questionably portable escape codes and
doesn't do many checks in terms of compatibility. It does try to see
if the terminal responds to commands, but doesn't currently do anything
with that information. More work could be done here in terms of
portability, but its a start.

Also fixed a few bugs that cropped up in the table implementation.